+++ CHANGELOG.md
@@ -6,12 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0.0] - unreleased
+## [2.0.0] - 2022-04-15
 
 ### Changed
 
+- _breaking_ `@PolymorphicResolver` returns a propertyDecorator
 - _breaking_ remove JSONArray for JSONValue[]
-- _breaking_ fromJSON only accepts `JSONValue`
+- _breaking_ fromJSON accepts `string | JSONObject`
 - _breaking_ fix #125 strategies are now passed the full array, this allows more
   flexibility for conversions
 - _breaking_ removes array as a strategy, use `composeStrategy` instead
@@ -34,6 +35,7 @@ and this project adheres to
 - update release node version to 16
 - std@0.133.0
 - Remove node tests
+- update Readme
 
 ## [v1.3.1] - 2021-03-14